+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_ldap_client.py
@@ -164,10 +164,33 @@ options:
     type: str
     version_added: 23.4.0
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
 notes:
   - LDAP client created using ZAPI should be deleted using ZAPI.
   - LDAP client created using REST should be deleted using REST.
   - REST only supports create, modify and delete data svm ldap client configuration.
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 
 '''
 
@@ -242,11 +265,14 @@ class NetAppOntapLDAPClient:
             user_dn=dict(required=False, type='str'),
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
             required_if=[
                 ('state', 'present', ['schema']),
+                ('use_lambda', True, ['lambda_config']),
             ],
             mutually_exclusive=[
                 ['servers', 'ad_domain'],
@@ -264,6 +290,8 @@ class NetAppOntapLDAPClient:
                                                ['group_dn', (9, 9, 0)], ['user_dn', (9, 9, 0)]]
         self.use_rest = self.rest_api.is_rest_supported_properties(self.parameters, unsupported_rest_properties, partially_supported_rest_properties)
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])