+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_quotas.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp, Inc
+# (c) 2018-2026, NetApp, Inc
 # GNU General Public License v3.0+
 # (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 from __future__ import absolute_import, division, print_function
@@ -114,6 +114,30 @@ options:
     choices: ['resize', 'reinitialize', 'none']
     type: str
     version_added: 20.12.0
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 
 '''
 
@@ -284,6 +308,8 @@ class NetAppONTAPQuotas:
             activate_quota_on_change=dict(required=False, type='str', choices=['resize', 'reinitialize', 'none'])
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
@@ -296,7 +322,10 @@ class NetAppONTAPQuotas:
                 'soft_disk_limit': ['quota_target', 'type'],
                 'threshold': ['quota_target', 'type'],
             },
-            required_together=[('quota_target', 'type')]
+            required_together=[('quota_target', 'type')],
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
         )
 
         self.na_helper = NetAppModule()
@@ -311,6 +340,8 @@ class NetAppONTAPQuotas:
         self.validate_parameters_ZAPI_REST()
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])