+++ plugins/modules/na_ontap_qos_policy_group.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp, Inc
+# (c) 2018-2026, NetApp, Inc
 # GNU General Public License v3.0+
 # (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
@@ -9,6 +9,7 @@ na_ontap_qos_policy_group
 '''
 
 from __future__ import absolute_import, division, print_function
+
 __metaclass__ = type
 
 
@@ -161,6 +162,31 @@ options:
         required: false
         choices: ['used_space', 'allocated_space']
         version_added: 22.8.0
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -302,6 +328,8 @@ class NetAppOntapQosPolicyGroup:
             ))
         ))
 
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
@@ -313,6 +341,9 @@ class NetAppOntapQosPolicyGroup:
                 ['fixed_qos_options', 'adaptive_qos_options'],
                 ['is_shared', 'adaptive_qos_options'],
                 ['is_shared', 'fixed_qos_options']
+            ],
+            required_if=[
+                ('use_lambda', True, ['lambda_config'])
             ]
         )
         self.na_helper = NetAppModule()
@@ -335,6 +366,8 @@ class NetAppOntapQosPolicyGroup:
         self.uuid = None
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             if 'adaptive_qos_options' in self.parameters: